Save question category when cloning a question #2351

+++ main/exercise/question.class.php
@@ -509,23 +509,26 @@ abstract class Question
      * in this version, a question can only have 1 category
      * if category is 0, then question has no category then delete the category entry
      * @param int $categoryId
+     * @param int $courseId
      * @return bool
      *
      * @author Hubert Borderiou 12-10-2011
      */
-    public function saveCategory($categoryId)
+    public function saveCategory($categoryId, $courseId = 0)
     {
-        $courseId = api_get_course_int_id();
+        $courseId = empty($courseId) ? api_get_course_int_id() : (int) $courseId;
+
         if (empty($courseId)) {
             return false;
         }
+
         if ($categoryId <= 0) {
-            $this->deleteCategory();
+            $this->deleteCategory($courseId);
         } else {
             // update or add category for a question
             $table = Database::get_course_table(TABLE_QUIZ_QUESTION_REL_CATEGORY);
-            $categoryId = intval($categoryId);
-            $question_id = intval($this->id);
+            $categoryId = (int) $categoryId;
+            $question_id = (int) $this->id;
             $sql = "SELECT count(*) AS nb FROM $table
                     WHERE
                         question_id = $question_id AND
@@ -551,18 +554,25 @@ abstract class Question
 
     /**
      * @author hubert borderiou 12-10-2011
+     * @param int $courseId
      * delete any category entry for question id
      * delete the category for question
+     * @return bool
      */
-    public function deleteCategory()
+    public function deleteCategory($courseId = 0)
     {
+        $courseId = empty($courseId) ? api_get_course_int_id() : (int) $courseId;
         $table = Database::get_course_table(TABLE_QUIZ_QUESTION_REL_CATEGORY);
-        $question_id = intval($this->id);
+        $questionId = (int) $this->id;
+        if (empty($courseId) || empty($questionId)) {
+            return false;
+        }
         $sql = "DELETE FROM $table
                 WHERE
-                    question_id = $question_id AND
-                    c_id = ".api_get_course_int_id();
+                    question_id = $questionId AND
+                    c_id = ".$courseId;
         Database::query($sql);
+        return true;
     }
 
     /**
@@ -1409,17 +1419,17 @@ abstract class Question
      * Duplicates the question
      *
      * @author Olivier Brouckaert
-     * @param  array   $course_info Course info of the destination course
+     * @param  array   $courseInfo Course info of the destination course
      * @return false|string     ID of the new question
      */
-    public function duplicate($course_info = [])
+    public function duplicate($courseInfo = [])
     {
-        if (empty($course_info)) {
-            $course_info = $this->course;
-        } else {
-            $course_info = $course_info;
+        $courseInfo = empty($courseInfo) ? $this->course : $courseInfo;
+
+        if (empty($courseInfo)) {
+            return false;
         }
-        $TBL_QUESTIONS = Database::get_course_table(TABLE_QUIZ_QUESTION);
+        $questionTable = Database::get_course_table(TABLE_QUIZ_QUESTION);
         $TBL_QUESTION_OPTIONS = Database::get_course_table(TABLE_QUIZ_QUESTION_OPTION);
 
         $question = $this->question;
@@ -1427,24 +1437,25 @@ abstract class Question
         $weighting = $this->weighting;
         $position = $this->position;
         $type = $this->type;
-        $level = intval($this->level);
+        $level = (int) $this->level;
         $extra = $this->extra;
+        $categoryId = $this->category;
 
         // Using the same method used in the course copy to transform URLs
-        if ($this->course['id'] != $course_info['id']) {
+        if ($this->course['id'] != $courseInfo['id']) {
             $description = DocumentManager::replaceUrlWithNewCourseCode(
                 $description,
                 $this->course['code'],
-                $course_info['id']
+                $courseInfo['id']
             );
             $question = DocumentManager::replaceUrlWithNewCourseCode(
                 $question,
                 $this->course['code'],
-                $course_info['id']
+                $courseInfo['id']
             );
         }
 
-        $course_id = $course_info['real_id'];
+        $course_id = $courseInfo['real_id'];
 
         // Read the source options
         $options = self::readQuestionOption($this->id, $this->course['real_id']);
@@ -1460,10 +1471,10 @@ abstract class Question
             'level' => $level,
             'extra' => $extra
         ];
-        $newQuestionId = Database::insert($TBL_QUESTIONS, $params);
+        $newQuestionId = Database::insert($questionTable, $params);
 
         if ($newQuestionId) {
-            $sql = "UPDATE $TBL_QUESTIONS 
+            $sql = "UPDATE $questionTable 
                     SET id = iid
                     WHERE iid = $newQuestionId";
             Database::query($sql);
@@ -1486,7 +1497,7 @@ abstract class Question
             }
 
             // Duplicates the picture of the hotspot
-            $this->exportPicture($newQuestionId, $course_info);
+            $this->exportPicture($newQuestionId, $courseInfo);
         }
 
         return $newQuestionId;
